Remove an unnecessary class
